Migrates react table usages & creates more seeders & factories
