+++ database/factories/LeaseFactory.php
@@ -3,6 +3,7 @@
 namespace Database\Factories;
 
 use App\Enums\Status;
+use App\Models\Unit;
 use App\Models\User;
 use Illuminate\Database\Eloquent\Factories\Factory;
 
@@ -19,9 +20,8 @@ class LeaseFactory extends Factory
     public function definition(): array
     {
         return [
+            "unit_id"     => Unit::factory(),
             "user_id"     => User::factory(),
-            "rent_amount" => $this->faker->numberBetween(20000, 100000),
-            "deposit"     => $this->faker->numberBetween(0, 50000),
             "expires_at"  => $this->faker->dateTimeBetween("+2 months", "+1 year"),
             "status"      => $this->faker->randomElement([Status::ACTIVE, Status::INACTIVE]),
             "created_at"  => $this->faker->dateTimeBetween('-1 years')
