+++ database/factories/PaymentPlanFactory.php
@@ -17,7 +17,9 @@ class PaymentPlanFactory extends Factory
     public function definition()
     {
         return [
-            //
+            "rent_amount" => $this->faker->numberBetween(20000, 100000),
+            "deposit"     => $this->faker->numberBetween(0, 50000),
+            "due_day"     => $this->faker->numberBetween(1, 31),
         ];
     }
 }
