+++ database/seeders/DatabaseSeeder.php
@@ -14,13 +14,14 @@ class DatabaseSeeder extends Seeder
     public function run(): void
     {
         $this->call([
-            RoleSeeder::class,
-            UserSeeder::class,
+//            RoleSeeder::class,
+//            UserSeeder::class,
 //            EstateSeeder::class,
 //            PropertySeeder::class,
 //            UnitSeeder::class,
+//            LeaseSeeder::class,
 //            ServiceSeeder::class,
-//            AmenitySeeder::class,
+            AmenitySeeder::class,
 //            ServiceProviderSeeder::class,
 //            TransactionSeeder::class,
 //            PaymentSeeder::class,
