+++ database/seeders/LeaseSeeder.php
@@ -2,6 +2,7 @@
 
 namespace Database\Seeders;
 
+use App\Models\Lease;
 use Illuminate\Database\Seeder;
 
 class LeaseSeeder extends Seeder
@@ -11,8 +12,9 @@ class LeaseSeeder extends Seeder
      *
      * @return void
      */
-    public function run()
+    public function run(): void
     {
-        //
+        Lease::factory(5)->create();
+        Lease::factory(3)->hasPaymentPlans(mt_rand(1, 3))->create();
     }
 }
